samples/a365/getting_started/pyproject.toml
[project]
name = "getting-started-with-distro-exporter"
version = "0.1.0"
description = "Sample demonstrating the Microsoft OpenTelemetry Distro with A365 exporter and span scopes."
authors = [{ name = "Microsoft", email = "example@microsoft.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "microsoft-opentelemetry",
    "aiohttp",
    "openai",
    "microsoft-agents-hosting-aiohttp>=0.5.0",
    "microsoft-agents-hosting-core>=0.5.0",
    "microsoft-agents-authentication-msal>=0.5.0",
    "microsoft-agents-activity>=0.5.0",
    "uvicorn[standard]~=0.35.0",
    "starlette>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "ruff>=0.6.0"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = ["__pycache__"]
